alterações para header mostrar LogOut
