+++ Frontend/src/app/app.component.html
@@ -3,7 +3,7 @@
 
     <div class="wrapper">
   
-      <app-header></app-header>
+      <app-header [user]="userLogged" ></app-header>
       
         <!-- Full Width Column -->
         <div class="content-wrapper">
